refactor(tiledit): massive refactor to split out shared library
- I don't want to need a QT dependency when running unit tests etc. and it might be handy for use in other tools in the toolchain
