+++ .github/workflows/sirc-tiledit.yml
@@ -1,5 +1,6 @@
 name: SIRC Tiledit
 
+# TODO: Only run when sirc-tiledit directory has changes
 on:
   push:
     branches: ["main"]
