+++ .vscode/launch.json
@@ -1,0 +1,17 @@
+{
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "lldb",
+      "request": "launch",
+      "name": "Debug sirc-tiledit-gui",
+      "program": "${workspaceFolder}/sirc-tiledit/build/sirc-tiledit-gui",
+      "args": [],
+      "cwd": "${workspaceFolder}/sirc-tiledit/build/"
+      // TODO: Compile before running
+    }
+  ]
+}
