+++ sirc-tiledit/.tool-versions
@@ -1,0 +1,2 @@
+meson 1.4.1
+ninja 1.11.1
