+++ sirc-tiledit/data/images/source/dusk/far-clouds.png
