+++ sirc-tiledit/data/images/source/dusk/far-mountains.png
