+++ sirc-tiledit/data/images/source/dusk/mountains.png
