+++ sirc-tiledit/data/images/source/dusk/near-clouds.png
