+++ sirc-tiledit/data/images/source/dusk/sky.png
