+++ sirc-tiledit/data/images/source/public-license.txt
