+++ sirc-tiledit/libs/gui/src/aboutdialog.hpp
