+++ sirc-tiledit/libs/gui/src/main.cpp
@@ -1,4 +1,4 @@
-#include <mainwindow.h>
+#include "mainwindow.hpp"
 
 #include <QApplication>
