+++ sirc-tiledit/libs/gui/ui/mainwindow.ui
