+++ sirc-tiledit/libs/shared/include/sircimage.hpp
@@ -1,9 +1,8 @@
 #ifndef IMAGEPROCESSOR_H
 #define IMAGEPROCESSOR_H
 
-#include <QImage>
-#include <QPixmap>
 #include <array>
+#include <cstddef>
 #include <cstdint>
 #include <map>
 #include <vector>
@@ -13,21 +12,21 @@ const int HEIGHT_PIXELS = 256;
 // The number of palette slots in the SIRC PPU
 const int MAX_PALETTE_SIZE = 256;
 
-// QColor uses standard 32 bit colour ARGB (8bpp)
-const unsigned int Q_COLOR_RANGE = 0xFF;
 // SIRC uses a packed 16 bit color RGB (5bpp)
 const unsigned int SIRC_COLOR_COMPONENT_BITS = 5;
 const unsigned int SIRC_COLOR_RANGE = (1 << (SIRC_COLOR_COMPONENT_BITS)) - 1;
-const unsigned int Q_TO_SIRC_COLOR_RATIO = Q_COLOR_RANGE / SIRC_COLOR_RANGE;
 
 using SircColor = uint16_t;
+using ArgbColor = uint32_t;
 using PaletteReference = uint8_t;
-using PixelData =
+using PackedPixelData =
+    std::array<std::array<SircColor, HEIGHT_PIXELS>, WIDTH_PIXELS>;
+using IndexedPixelData =
     std::array<std::array<PaletteReference, HEIGHT_PIXELS>, WIDTH_PIXELS>;
 
 struct SircImageData {
   std::vector<SircColor> palette;
-  PixelData pixelData;
+  IndexedPixelData pixelData;
 };
 
 /**
@@ -40,12 +39,8 @@ struct SircImageData {
 class SircImage {
 
 public:
-  static SircImage fromQPixmap(const QPixmap &pixmap);
+  static SircImage fromPixelData(const PackedPixelData &pixelData);
   static SircImage fromSircImageData(const SircImageData &imageData);
-
-  // TODO: Avoid doing palette reduction in each of the following functions
-  [[nodiscard]] QPixmap toQPixmap() const;
-  [[nodiscard]] std::vector<QColor> getPaletteColors() const;
   [[nodiscard]] SircImageData getImageData() const;
 
 private:
